Serval new types have been added

+++ include/types/Birthdate.hxx
@@ -10,8 +10,8 @@ namespace telegram {
             public:
                 typedef std::shared_ptr<Birthdate> Ptr;
 
-                std::uint8_t day;
-                std::uint8_t month;
+                std::uint16_t day;
+                std::uint16_t month;
                 std::uint32_t year;
             };
         }
